requirements.txt
agno==1.1.3
streamlit
pydantic
python-docx
prefect